fix[MODEL]: Update category model to return an array for CO2eq
CO2eqs can return different CO2eq values with different units
